+++ server/src/main/java/com/codueon/boostUp/domain/lesson/dto/Get/GetSearchLesson.java
@@ -1,4 +1,4 @@
-package com.codueon.boostUp.domain.lesson.dto;
+package com.codueon.boostUp.domain.lesson.dto.Get;
 
 import lombok.Builder;
 import lombok.Getter;
