+++ server/src/main/java/com/codueon/boostUp/domain/lesson/dto/Patch/PatchLessonCurriculum.java
@@ -1,4 +1,4 @@
-package com.codueon.boostUp.domain.lesson.dto;
+package com.codueon.boostUp.domain.lesson.dto.Patch;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,11 +6,11 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class GetLessonCurriculum {
+public class PatchLessonCurriculum {
     private String curriculum;
 
     @Builder
-    public GetLessonCurriculum(String curriculum) {
+    public PatchLessonCurriculum(String curriculum) {
         this.curriculum = curriculum;
     }
 }
