+++ server/src/main/java/com/codueon/boostUp/domain/lesson/dto/Post/PostLesson.java
@@ -1,4 +1,4 @@
-package com.codueon.boostUp.domain.lesson.dto;
+package com.codueon.boostUp.domain.lesson.dto.Post;
 
 import lombok.Builder;
 import lombok.Getter;
